Response ClientResponse.read() override
This does not play that nicely with the upstream aiohttp changes, and
the internal API has changed (e.g., _data -> _body) in between supported
releases.

Looking at where read() is added, it's not clear to me whether this is
still likely to be a problem with current aiohttp, but if it does turn
out to cause issues then it's probably worth addressing there instead of
using these internal attributes.