aiohttp does not accept generators for streaming data
The aiohttp FormData api expects the given data to be either a file like
object, multidict or (name, file) pairs, not a generator. The aiohttp
documention mentions streaming uploads using asynchronous generators
here:
https://docs.aiohttp.org/en/stable/client_quickstart.html#streaming-uploads

However we've decided against it, since the documents we're bulk
updating are passed and in memory already. Additionally, we expect
sending < 100 documents per bulk update and no funding is available for
maintenance at this point. Given these constraints, we decided to remove
the generator and simply pass the docs along to the API.

The new_edits parameter defaults to 'True' to reflect the CouchDB >= 2
API's defaults.

Tests have been adjusted to reflect this. One test testing the
all_or_nothing parameter has been deleted since the CouchDB API has
removed support for this in CouchDB >= 2